chore : binarySearch

+++ c_code/iconics.c
@@ -4,9 +4,9 @@
 
 // Binary search
 
-int binary_search(int arr[], int n, int target) {
-    int low = 0, high = n - 1;
-    while (low <= high) {
+int binary_search(int arr[], int n,  int target) {
+    unsigned int low = 0, high = n - 1;
+    while ((int)low <= (int)high) {
         int mid = low + (high - low) / 2;
         if (arr[mid] == target) {
             return mid;
